# make the function faster than the before