--- a/code/lorentz.lisp
+++ b/code/lorentz.lisp
@@ -1,29 +1,22 @@
-(defparameter dt 0.01d0)
-(defparameter n 2000000)
-(defparameter a 10.0d0)
-(defparameter b 28.0d0)
-(defparameter c (/ 8.0d0 3))
+(defconstant dt 0.01d0)
+(defconstant n 2000000)
+(defconstant a 10.0d0)
+(defconstant b 28.0d0)
+(defconstant c #.(/ 8.0d0 3)) 
 
-(defvar x (make-array (list (+ n 1)) :initial-element 0.0d0))
-(defvar y (make-array (list (+ n 1)) :initial-element 0.0d0))
-(defvar z (make-array (list (+ n 1)) :initial-element 0.0d0))
+(defun main ()
+  (declare (optimize (debug 0) (safety 0) (speed 3)))
+  (let ((x (make-array n :element-type 'double-float))
+	(y (make-array n :element-type 'double-float))
+	(z (make-array n :element-type 'double-float)))
+    (setf (aref x 0) 0.0d0
+	  (aref y 0) 1.0d0
+	  (aref z 0) 1.05d0)
+    (dotimes (i #.(1- n))
+      (declare (fixnum i))
+      (setf (aref x (1+ i)) (+ (aref x i) (* #.dt (- (* #.a (aref y i)) (* #.a (aref x i)))))
+	    (aref y (1+ i)) (+ (aref y i) (* #.dt (- (* #.b (aref x i)) (* (aref x i) (aref z i)) (aref y i))))
+	    (aref z (1+ i)) (+ (aref z i) (* #.dt (- (* (aref x i) (aref y i)) (* #.c (aref z i)))))))
+    ))
 
-
-(setf (aref x 1) 0.0d0)
-(setf (aref y 1) 1.0d0)
-(setf (aref z 1) 1.05d0)
-
-(time
- (loop for i from 1 to (1- n)
-    do 
-      (setf (aref x (1+ i))
-            (+ (aref x i)
-               (* dt (+ (- (* a (aref x i))) (* a (aref y i))))))
-      (setf (aref y (1+ i))
-            (+ (aref y i)
-               (* dt (+ (- (* (aref x i) (aref z i))) (* b (aref x i))  (- (aref y i))))))
-      (setf (aref z (1+ i))
-            (+ (aref z i)
-               (* dt (- (* (aref x i) (aref y i)) (* c (aref z i))))))))
-
-
+(time (main))