以前のコードに以下の最適化を施したところ34倍高速になりましたので報告します。

- 定数化
- 定数のインライン展開
- 配列の型指定
- 大域変数の除去

これでJuliaの最高速との差が３倍以内になりましたね。

- オリジナル

```
Evaluation took:
  1.546 seconds of real time
  1.539931 seconds of total run time (1.321762 user, 0.218169 system)
  [ Run times consist of 0.793 seconds GC time, and 0.747 seconds non-GC time. ]
  99.61% CPU
  2,783,102,624 processor cycles
  607,985,680 bytes consed
```
- 最適化後

```
Evaluation took:
  0.045 seconds of real time
  0.044398 seconds of total run time (0.026790 user, 0.017608 system)
  97.78% CPU
  80,243,036 processor cycles
  48,000,048 bytes consed
```